+++ tests/testthat/test-Model.R
@@ -1,7 +1,6 @@
 test_that("Model initialisation", {
   m <- Model$new(list(x = 1), name = "model")
   expect_true(m$name == "model")
-
   m <- Model$new(list(x = 1))
   expect_null(m$null)
 })